feat: 에뮬레이터 홈 화면 구현
