+++ monicar-emulator/build.gradle
@@ -1,5 +1,6 @@
 dependencies {
     implementation "org.springframework.boot:spring-boot-starter-web"
+    implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
     implementation "org.springframework.boot:spring-boot-starter-validation"
 
     implementation 'org.springframework:spring-aspects'
